# perf(format): add fast path using json stringify/parse