@H4ad 2024-01-13
Before merging, we should also take a look a compare this change with #81 and #82, which also improves the performance but in a more general way than this fast path.