@aeschli 2024-01-22
The point of the formatter is to provide edits that describe the minimal code changes.
A client can always use JSON.stringify(JSON.parse(..) ) if that is not needed.